Crate rust_odata

Rust oData

Rust oData is library for building oData REST services in Rust. This library is referenced and utilized by the Connect Four repository for client-server communication.

Resources are declared using the Entity Data Model

Implementation Status

For the project, likely only a subset of oData 4.0 will be implemented, with more advanced features potentially implemented later or for the project if time allows. The focus here is on core functionality.

Current status (0-3)

Entity Class - 1

Entity Set Class - 2

Edm Types - 2

Service Class - 0

Model registration & hookup

Validation

Create request

Read request (by ID, by property, by value)

Update request

Delete request

Error codes

Metadata document - 0

Relationship Operations - 0

Singleton requests - 0

Singleton property requests -0

Contained Entities - 0

Potential additional features

Query & parameters

filter

expand

top

limit

count

select

search

lambda operators

Functions/actions

Etag modifcation/delete requests

Derived entity types

Open Entity/Complex Types

Batch requests

Interface

The original idea was to have users declare derived classes of entity/entity set respectively, but because Rust does not have C++ style inheritence, we are currently in the process of rethinking this.

One idea is to instead use the Rust builder paradigm to specify entity types, ie.

let product = EntityBuilder::new("Product")
    .add("id",   Edm::Type::String)
    .add("name", Edm::Type::String)
    .build();

where each .add() creates a new property for the specified Entity. This could also be expanded to include a validation policies:

let policy = Validation::new(|x| (x < 10) && (x > 0));
validation.add(|x| x mod 2);
product.add("age", Edm::Type::Int32, policy);

EntitySet could then be a trait, with a declare function used to specify the entity type to which it refers. CRUD-Q methods would then optionally be defined, with unimplemented stubs defaulting to a function which returns false denoting no implementation (perhaps this should instead return an error type? This would allow stubs to specify some kind of error and return it).

impl EntitySet for TestProducts {

    fn declare (&self) -> Entity {
        let product = EntityBuilder::new("Product")
            .add("id",   Edm::Type::String)
            .add("name", Edm::Type::String)
            .build();
        product
    }

    fn create (&self, data: JSON<or generic obj?>) -> bool {
        // do something with input ...
        true
    }

    fn read ...
}

This is all very tenative though! Currently exploring macros as a way to potentially improve interface. Right now specifiy properties in this fashion appears to require the use of a HashMap. This is less than ideal as we give up quite a bit of performance than we would otherwise have by using native Rust structs.

Modules

edm

entity

entity_set

service